backend/requirements.txt
fastapi>=0.109.1
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
langgraph>=0.0.26
langchain>=0.1.7
langchain-core>=0.1.23
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
pyyaml>=6.0.1
apscheduler>=3.10.4
python-dotenv>=1.0.0
